added heartbeat

+++ include/gpio_stm32f0.h
@@ -7,6 +7,6 @@
 enum Mode {INPUT, OUTPUT, ALTERNATE, ANALOGUE};
 
 void clock_enable(GPIO_TypeDef *port);
-void set_mode(GPIO_TypeDef *port, uint32_t pin, uint8_t mode);
+void set_mode(GPIO_TypeDef *port, uint16_t pin, uint8_t mode);
 
 #endif
