+++ src/gpio_stm32f0.c
@@ -17,7 +17,7 @@ void clock_enable(GPIO_TypeDef *port)
 }
 
 
-void set_mode(GPIO_TypeDef *port, uint32_t pin, uint8_t mode)
+void set_mode(GPIO_TypeDef *port, uint16_t pin, uint8_t mode)
 {
 	switch (mode)
 	{
